Properly check for raid level 0
The default attributes for the raid level is the integer 0. The check
for the raid level is looking for the string "0". This causes a bug
where a configuration with raid level 0 will go to the else case and
fail because bitmap is passed as an argument to the mdadm command.